+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/table/DefaultFilterTableModelTest.java
@@ -208,7 +208,7 @@ public final class DefaultFilterTableModelTest {
 										.build();
 		testModel.selection().indexes().addListener(selectionEvents::incrementAndGet);
 		testModel.items().refreshStrategy().set(RefreshStrategy.MERGE);
-		testModel.sorter().sort(0).set(SortOrder.ASCENDING);
+		testModel.sorter().order(0).set(SortOrder.ASCENDING);
 		testModel.items().refresh();
 		testModel.selection().index().set(1);//b