+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTable.java
@@ -739,13 +739,13 @@ public final class FilterTable<R, C> extends JTable {
 	private void toggleColumnSorting(int selectedColumn, boolean add) {
 		if (sortingEnabled.get() && selectedColumn != -1) {
 			C identifier = columnModel().getColumn(selectedColumn).identifier();
-			if (!tableModel.sorter().sort(identifier).locked().get()) {
+			if (!tableModel.sorter().order(identifier).locked().get()) {
 				ColumnSortOrder<C> columnSortOrder = tableModel.sorter().columnSort().get(identifier);
 				if (add) {
-					tableModel.sorter().sort(identifier).add(nextSortOrder(columnSortOrder.sortOrder()));
+					tableModel.sorter().order(identifier).add(nextSortOrder(columnSortOrder.sortOrder()));
 				}
 				else {
-					tableModel.sorter().sort(identifier).set(nextSortOrder(columnSortOrder.sortOrder()));
+					tableModel.sorter().order(identifier).set(nextSortOrder(columnSortOrder.sortOrder()));
 				}
 			}
 		}
@@ -982,16 +982,16 @@ public final class FilterTable<R, C> extends JTable {
 			int index = columnModel.getColumnIndexAtX(e.getX());
 			if (index >= 0) {
 				C identifier = columnModel.getColumn(index).identifier();
-				if (!tableModel.sorter().sort(identifier).locked().get()) {
+				if (!tableModel.sorter().order(identifier).locked().get()) {
 					if (!getSelectionModel().isSelectionEmpty()) {
 						setColumnSelectionInterval(index, index);//otherwise, the focus jumps to the selected column after sorting
 					}
 					ColumnSortOrder<C> columnSortOrder = tableModel.sorter().columnSort().get(identifier);
 					if (e.isAltDown()) {
-						tableModel.sorter().sort(identifier).add(nextSortOrder(columnSortOrder.sortOrder()));
+						tableModel.sorter().order(identifier).add(nextSortOrder(columnSortOrder.sortOrder()));
 					}
 					else {
-						tableModel.sorter().sort(identifier).set(nextSortOrder(columnSortOrder.sortOrder()));
+						tableModel.sorter().order(identifier).set(nextSortOrder(columnSortOrder.sortOrder()));
 					}
 				}
 			}